+++ tailwind.config.js
@@ -1,3 +1,8 @@
+import {
+  scopedPreflightStyles,
+  isolateInsideOfContainer, // there are also isolateOutsideOfContainer and isolateForComponents
+} from "tailwindcss-scoped-preflight";
+
 const REM_FACTOR = 0.25;
 
 const spacing = Array.from(Array(1001).keys()).reduce(
@@ -9,6 +14,9 @@ const spacing = Array.from(Array(1001).keys()).reduce(
 export default {
   content: ["./src/**/*.tsx", "./src/**/*.ts"],
   important: true,
+  corePlugins: {
+    preflight: false,
+  },
   theme: {
     extend: {
       maxWidth: ({ theme }) => ({ ...theme("spacing"), "8xl": "90rem" }),